requirements.txt
# Core dependencies
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.0
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1

# CLI et utilitaires
click==8.1.7
rich==13.7.0

# Tests
pytest==7.4.3
pytest-cov==4.1.0

# Développement
black==23.12.0
flake8==6.1.0
pandas>=2.0.0
    